requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pillow==10.4.0
torch==2.3.0
transformers==4.41.0
requests==2.32.3
python-multipart==0.0.9
google-generativeai==0.5.4
python-dotenv==1.0.1
